Add draft of VtkToAlembic application
This code was designed to take the shape two black holes' event horizons
as they merged (as simulated with SpEC and post-processed in ParaView)
and deliver it to VFX studios in a format they could use for artistic
rendering.  At the time, Alembic wasn't particularly well-documented, and
there was no support in Blender.  The code is therefore very
special-purpose, containing hard-coded paths to directories, baked-in
assumptions about field names, etc.  But hopefully it's at least a good
reference for using the VTK and Alembic APIs.

As configured, it built on Ubuntu back in early 2016 with user installs
of VTK, ilmbase, and alembic.  CMake was used solely to manage the VTK
dependency nightmare.

Relates to #3.